requirements.txt
Flask
python-dotenv
supabase
Flask-Cors
Werkzeug